backend/requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0

